requirements.txt
streamlit==1.0.0
python-dotenv==0.19.1
azure-ai-formrecognizer==3.1.2
azure-core==1.19.0
azure-storage-blob==12.9.0
langchain==0.1.0  # Replace with actual version
